Add files via upload
t¬¬ 

Table of Contents
Project Idea:	3
1)	Introduction:	3
2)	Background:	3
Project Scope:	4
Technology to be used:	5
Functional Requirements:	5
Basic Flow of the Program:	6
Flow Chart:	6
Objectives:	7
Intended Audience:	7
Team Dynamics:	7
1)	Group Members:	7
2)	Workload Distributions:	7
Road Map:	8
Conclusion:	9
References:	9

 
Project Idea:
1)	Introduction:
A game is a systematic kind of fun activity that is usually done for pleasure but can also be used for educational purposes. Games are distinct from employment, which is typically done for a financial benefit, and art, which is more commonly an expression of creativity. However, the line between work and play is blurry, and many games are also considered work (for example, professional spectator sports or e-sports gamers).
	Games are played for a variety of reasons, including pure fun, achievement, fitness, and reward. They can be played by enthusiasts or pros, and they can be played alone, in teams, or online.
Goals, rules, challenges, and engagement are all important aspects of games. Games generally involve some form of mental or physical stimulation, and in some cases, both. Many games help in the development of practical skills, serve as a kind of exercise, or have an educational, simulation, or psychological function.
	Games are of different types some are played on the field, table games, and video games (Avedon & Sutton-Smith, n.d.)
Physical Games	Virtual Games
Cricket	Cricket 2007,2020 and many  more
Football	FIFA
Tennis	Virtua Tennis, Tennis World Tour
Chess	Chess Ultra
Maze Solving	Pac-Man

2)	Background:
The player controls Pac-Man with either keyboard arrows or a joystick to begin the game. The objective is to guide Pac-Man through the maze-like screen while avoiding or assaulting one of four chasing ghosts (sometimes called monsters).
Blinky (red), Inky (light blue), Pinky (pink), and Clyde (green) are the four ghosts (orange). Each ghost has a unique offensive strategy: Blinky, for example, is known as Shadow because it moves the fastest. The ghosts eventually leave the "ghost prison" at the maze's center and roam the board as the game proceeds. Pac-Man loses a live and the game resets if he collides with a ghost.
       Each level has four power pellets available in the corners. (Wikipedia, n.d.)
Project Scope:
The original Pac-man game was in high graphics and had only one level. In the original Pac-Man, the player must eat all grains (dots) on the map to win the game. 5 special grains activate the power to eat the ghosts and when the ghost dies it respawns in the cage and gets stuck for a few seconds. 
 
Our version of the Pac-Man is console-based and has low graphics. As in the original game, there are several grains, but the player doesn’t have to eat all the grains to win the game. But if we must win the game, we must eat all the special grains. There are also some riddles in our game if we solve them, we can win the game in less time. In our game, if we hit the boundary (wall), the boundary gets destroyed but we lost lives in the original version everything remains the same. (Ke Xu, 2004)

 
Technology to be used:
•	Visual Studio (Console)
•	Language C++
Functional Requirements:
1)	Maps:
1.1)	 According to levels:
	In our Pac-Man version game, we add different maps according to levels if the player pass level 1 then moves to level 2, and so on.
1.2)	According to difficulty:
	Player have the choice to choose the difficulty level according to his/her choice (i.e., Normal, Medium, Hard)

2)	Scoring System:
	Our game is based upon a point system.
	Grain scoring system.
	In a game, if the player eats up all number of grains in the game or eats the bigger circular grain, then the player wins or moves to the next level. 
	The small grains are worth 1-point, bigger grains are worth 5-points, and the next-level train is worth 10-points.

3)	Levels:
If we complete the level the player moves to the next level and its keep moving to the next level until live gets zero or all the level get completed. 
4)	Choose Maps:
Player have the choice to choose the map individually according to his/her choice.
5)	Lives:
There are 3 lives if the player hit the ghost the live gets minus and if the player loses all 3 live the game gets over. 
6)	Leaderboard:
When the level is finished, the score of all the levels is collected and saved in the text file, the player can see the score of the other on the leader board.
7)	Game-Over:
If the player loses all 3 lives, the game gets over. 
 
Basic Flow of the Program:
	In our version of the at the beginning of the game, we will have the main menu from where we can select different modes Level mode, Difficulty mode, and Map mode. In the end, the total will display which will store in the leaderboard section.

Flow Chart:




Objectives:
•	Entertainment 
•	Mind Relaxation 
•	Strategy building

Playing games regularly can improve children's short and long-term memory, as well as help the brain digest information more quickly. In addition, games engage players' imaginations, allowing them to stay focused on certain tasks while also increasing their patience in achieving a goal.
Our main objective is to entertain people with our game and when they are stressed, they can play this game to relax and refresh their minds. It can also help in strategy building.  
Intended Audience:
The intended audience for this game is mainly adults and broadly anyone can play this.   
Conclusion:
	We believe the project was successfully finished. The game is enjoyable to play. We never expected Pac-Man would be so difficult and time-consuming to create. We learned a lot at that time, and we'd want to thank our teacher Mam, Nabia Khalid- Lecturer (Computer Science) at Bahria University, Islamabad. We would not have been able to finish the job without her assistance and dedication. She made a significant contribution to our learning.
References:
•	Retrieved from Wikipedia: https://en.wikipedia.org/wiki/Pac-Man
•	Avedon, E., & Sutton-Smith, B. (n.d.). Wikipedia.
Retrieved from https://en.wikipedia.org/wiki/Game
•	Ke Xu, E. L. (2004).
Retrieved from http://www1.cs.columbia.edu/~sedwards/classes/2004/4840/reports/pacman.pdf